Conflicts:
	fannie/modules/plugins2.0/HouseCoupon/HouseCouponEditor.php
	fannie/modules/plugins2.0/HouseCoupon/index.php
	pos/is4c-nf/graphics/is4c.gif
	pos/is4c-nf/gui-modules/nslogin.php
	pos/is4c-nf/install/extra_config.php
	pos/is4c-nf/install/index.php
	pos/is4c-nf/lib/Database.php
	pos/is4c-nf/lib/PrehLib.php
	pos/is4c-nf/lib/SQLManager.php
	pos/is4c-nf/lib/Scanning/SpecialUPCs/HouseCoupon.php
	pos/is4c-nf/lib/TotalActions/ArOverpayAction.php
	pos/is4c-nf/plugins/QuickMenus/quickmenus/1.php

+++ fannie/install/sample_data/InstallSampleDataPage.php
@@ -133,9 +133,11 @@ elseif(isset($_REQUEST['memtype'])){
 	echo "Loading memtype";
 	$db->query("TRUNCATE TABLE memtype");
 	loaddata($db,'memtype');
-	echo "Loading memdefaults";
-	$db->query("TRUNCATE TABLE memdefaults");
-	loaddata($db,'memdefaults');
+    if ($db->tableExists('memdefaults')) {
+        echo "Loading memdefaults";
+        $db->query("TRUNCATE TABLE memdefaults");
+        loaddata($db,'memdefaults');
+    }
 }
 elseif(isset($_REQUEST['products'])){
 	echo "Loading products";
@@ -274,8 +276,6 @@ Load default place-of-origin information<br />
 // InstallSampleDataPage
 }
 
-if (basename(__FILE__) == basename($_SERVER['PHP_SELF'])){
-	$obj = new InstallSampleDataPage();
-	$obj->draw_page();
-}
+FannieDispatch::conditionalExec(false);
+
 ?>
